requirements.txt
# Core LLM providers
openai>=1.30.0
anthropic>=0.34.0
google-genai>=0.3.0

# HTTP & async
httpx>=0.27.0
requests>=2.32.0
tenacity>=8.0.0

# Configuration & environment
python-dotenv>=1.0.1
pyyaml>=6.0
toml>=0.10.2

# Data validation
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0

# Utilities
loguru>=0.7.0
tqdm>=4.65.0
deepdiff>=6.0.0
addict>=2.4.0
